% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBN.R
\name{DBN}
\alias{DBN}
\title{Deep Belief Network}
\usage{
DBN(x, y, n.iter = 300, nodes = c(30, 40, 30), learning.rate = 0.5,
  size.minibatch = 10, n.iter.pre = 30, learning.rate.pre = 0.1,
  verbose = FALSE)
}
\arguments{
\item{x}{A matrix with binary features of shape samples * features.}

\item{y}{A matrix with labels for the data. (Always needs to be provided for training the DBN)}

\item{n.iter}{The number of epochs to run backpropogation.}

\item{nodes}{A vector with the number of hidden nodes at each layer}

\item{learning.rate}{Learning rate for supervised finetuning of Stacked RBM.}

\item{size.minibatch}{The size of the minibatches used for training.}

\item{n.iter.pre}{Past on to the StackRBM function, defines the how many epochs are used to pretrain each
RBM layer.}

\item{learning.rate.pre}{The pretraining learning rate, passed on to the StackRBM function.}

\item{verbose}{Whether to print th training error at each epoch, printing will slow down the fitting.}
}
\value{
Returns the finetuned DBN model that can be used in the PredictDBN function.
}
\description{
Trains a deep belief network starting with a greedy pretrained stack of RBM's (unsupervised) using the function
StackRBM and then DBN adds a supervised output layer. The stacked RBM is then finetuned on the supervised criterion
by using backpropogation.
}
