% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{digits}
\alias{digits}
\title{Handwritten digit data from Kaggle}
\format{A matrix with 42,000 rows and 785 columns:}
\source{
\url{https://www.kaggle.com/c/digit-recognizer/data}
}
\usage{
digits
}
\description{
42000 digits with 1 class and 784 pixel intensities.
}
\keyword{datasets}

