% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.R
\name{down.rbm}
\alias{down.rbm}
\title{Reconstruct original data using features}
\usage{
down.rbm(model, data)
}
\arguments{
\item{model}{an RBM object from the single.rbm or multi.rbm functions}

\item{data}{matrix with features}
}
\value{
a matrix with reconstructions of the input features
}
\description{
Reconstruct original data using features
}
\examples{
# Data availible at https://www.samverkoelen.com/data
data <- as.matrix(read.csv('mnist_sample.csv'));

#train RBM stack
model <- multi.rbm (hidden = c(150, 75, 30), data = train)

par(mfrow=c(1,2))
id <- 57

#Original
original <- matrix(data[ ,id], byrow = T, ncol = 16)
original <- t(original[16:1, 1:16]) #correct orientation
image(original, col = grey(seq(0, 1, 0.001)), main = 'Original')

#visible to hidden
features <- up.rbm(model, data)

#hidden back to visible
reconstructions <- down.rbm(model, features)

#Reconstruction
reconstruction <- matrix(reconstructions[id, ], byrow = T, ncol = 16)
reconstruction <- t(reconstruction[16:1, 1:16]) #correct orientation
image(reconstruction, col = grey(seq(0, 1, 0.001)), main = 'Reconstruction')
}

