% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{george_reviews}
\alias{george_reviews}
\title{A single person's movie reviews}
\format{A Matrix with 1 rows and 6 columns:}
\source{
\url{https://github.com/echen/restricted-boltzmann-machines}
}
\usage{
george_reviews
}
\description{
1 reviews for 6 movies
}
\keyword{datasets}

