% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hill.R
\name{hill}
\alias{hill}
\title{Hill estimator}
\usage{
hill(data_array, idx = 1:(length(logsorted) - 1))
}
\arguments{
\item{data_array}{the raw data, assumed to be positive}

\item{idx}{Hill tuning parameter}
}
\description{
Hill estimator for tail index estimation, provided as a baseline
}

