% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{movie_reviews}
\alias{movie_reviews}
\title{Sample movie reviews}
\format{A Matrix with 6 rows and 6 columns:}
\source{
\url{https://github.com/echen/restricted-boltzmann-machines}
}
\usage{
movie_reviews
}
\description{
6 reviews for 6 movies
}
\keyword{datasets}

