% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.R
\name{multi.rbm}
\alias{multi.rbm}
\title{Train multiple layers of RBM's}
\usage{
multi.rbm(hidden, data, learning_rate = 0.1, epochs = 1000,
  batch_size = 100, momentum = 0.9, verbose = T)
}
\arguments{
\item{hidden}{a vector containing the number of hidden units per layer}

\item{data}{matrix with training data}

\item{learning_rate}{a numeric value spacifying the learning rate or a vector containing these values for each layer}

\item{epochs}{the number of iterations per batch or a vector containing these values for each layer}

\item{batch_size}{during one epoch a single batch will be trained. This number must be divisible by the number of training items}

\item{momentum}{a numeric value spacifying the momentum or a vector containing these values for each layer}

\item{verbose}{logical value for showing the current status in the console}
}
\value{
RBM object
}
\description{
Train multiple layers of RBM's
}
\examples{
# Data availible at https://www.samverkoelen.com/data
data <- as.matrix(read.csv('mnist_sample.csv'));

#train RBM stack
model <- multi.rbm (hidden = c(150, 75, 30), data = train)

par(mfrow=c(1,2))
id <- 57

#Original
original <- matrix(data[ ,id], byrow = T, ncol = 16)
original <- t(original[16:1, 1:16]) #correct orientation
image(original, col = grey(seq(0, 1, 0.001)), main = 'Original')

#visible to hidden
features <- up.rbm(model, data)

#hidden back to visible
reconstructions <- down.rbm(model, features)

#Reconstruction
reconstruction <- matrix(reconstructions[id, ], byrow = T, ncol = 16)
reconstruction <- t(reconstruction[16:1, 1:16]) #correct orientation
image(reconstruction, col = grey(seq(0, 1, 0.001)), main = 'Reconstruction')
}

