\name{ovarian_cancer_methylation}
\docType{data}
\alias{ovarian_cancer_methylation}
\title{ovarian cancer methylation example from United Kingdom Ovarian Cancer Population Study (UKOPS)}
\description{
This data set contains DNA methylation level from 1000 DNA methylation loci in 8 randomly selected women with 4 ovarian cancer cases (pre-treatment) and 4 age-matched healthy controls.
}
\usage{ovarian_cancer_methylation}
\value{
The ovarian cancer methylation example data set contains the following information:
 \item{IlmnID}{Name of DNA methylation loci}
 \item{case}{Ovarian cancer patients}
 \item{control}{Healthy controls}
}

\format{A matrix containing 1000 rows and 8 columns with each row denoting a methyaltion locus and each column denoting a subject.}
\source{NCBI GEO website with access number GSE19711}
\references{ 
Teschendorff AE, Menon U, Gentry-Maharaj A, Ramus SJ et al. Age-dependent DNA methylation of genes that are suppressed in stem cells is a hallmark of cancer. Genome Res 2010 Apr;20(4):440-6. PMID: 20219944
}
\keyword{datasets}