% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rbm.R
\name{plot.rbm}
\alias{plot.rbm}
\title{Plot method for a Restricted Boltzmann Machine}
\usage{
\method{plot}{rbm}(object, ...)
}
\arguments{
\item{...}{not used}

\item{x}{a RBM object}
}
\description{
This function plots the training error from an RBM
}

