% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictRBM.R
\name{PredictRBM}
\alias{PredictRBM}
\title{Predict RBM}
\usage{
PredictRBM(test, labels, model, layers)
}
\arguments{
\item{test}{Is the test-data (matrix) on which the user wants to make predictions.}

\item{labels}{Is a matrix with the corresponding labels for test-data.}

\item{model}{Is the trained RBM or StackRBM model.}

\item{layers}{Only needed with StackRBM to define the number of layers.}
}
\value{
A list with a confusion matrix and the accuracy of the predictions.
}
\description{
Function for predicting on a test set with either a classification RBM (trained with the RBM function) or a 
stacked RBM with a classification RBM on top (trained with StackRBM).
}
