% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rbm.R
\name{print.rbm}
\alias{print.rbm}
\title{Print method for a Restricted Boltzmann Machine}
\usage{
\method{print}{rbm}(object, ...)
}
\arguments{
\item{...}{not used}

\item{x}{a RBM object}
}
\description{
This function prints the weights for a RBM
}

