% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/random_sparse_matrix.R
\name{random_sparse_matrix}
\alias{random_sparse_matrix}
\title{Random Sparse Matrix}
\usage{
random_sparse_matrix(nrow, ncol, nnz, rfunc = rnorm, rand.x = function(n)
  round(rfunc(nnz), 2), ...)
}
\arguments{
\item{nrow}{number of rows in matrix}

\item{ncol}{number of cols in matrix}

\item{nnz}{number of non-zero entries}

\item{rfunc}{random function to use to generate the non-zero entries}

\item{rand.x}{random number generator for 'x' slot}

\item{...}{optionally further arguments passed to sparseMatrix()}
}
\value{
a sparseMatrix of dimension (nrow, ncol)
}
\description{
Random Sparse Matrix
}
\examples{
M1 <- random_sparse_matrix(1000, 20, nnz = 200)
summary(M1)
}
\author{
Martin Maechler, Zach Mayer
}
\references{
\itemize{
\item \link[Matrix]{sparseMatrix}
\item \url{http://www.inside-r.org/r-doc/Matrix/sparseMatrix}
}
}

