% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rbm.point_estimate.R
\name{rbm.point_estimate}
\alias{rbm.point_estimate}
\title{Get point estimate for Gamma}
\usage{
rbm.point_estimate(data_array)
}
\arguments{
\item{data_array}{the raw data, assumed to be positive}
}
\value{
A tuple: K -- the optimal tuning parameter chosen by RBM, Gamma -- the tail index estimate, Stdv -- a standard error estimate for Gamma
}
\description{
Uses RBM to get a point estimate for the
tail index Gamma.
}
\examples{
rbm.point_estimate(abs(rt(1000, df = 3)))
}

