% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconstructRBM.R
\name{ReconstructRBM}
\alias{ReconstructRBM}
\title{Reconstruct RBM}
\usage{
ReconstructRBM(test, model, layers)
}
\arguments{
\item{test}{A test example to reconstruct, works only with one example.}

\item{model}{A model trained with either the RBM or StackRBM function.}

\item{layers}{The number of layers that was used for training the StackRBM function.}
}
\value{
A plot of the original and reconstructed image (model) side by side.
}
\description{
Reconstructs an input image with a trained RBM or stacked RBM.
}
