% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.R
\name{single.rbm}
\alias{single.rbm}
\title{Train a single layer RBM}
\usage{
single.rbm(hidden, data, learning_rate = 0.1, epochs = 1000,
  batch_size = 100, momentum = 0.9, verbose = T, custom_verbose = "")
}
\arguments{
\item{hidden}{number of units in the hidden layer}

\item{data}{matrix with training data}

\item{learning_rate}{a numeric value spacifying the learning rate}

\item{epochs}{the number of iterations per batch}

\item{batch_size}{during one epoch a single batch will be trained. This number must be divisible by the number of training items}

\item{momentum}{a numeric value spacifying the momentum}

\item{verbose}{logical value for showing the current status in the console}
}
\value{
RBM object
}
\description{
Train a single layer RBM
}
\examples{
# Data availible at https://www.samverkoelen.com/data
data <- as.matrix(read.csv('mnist_sample.csv'));

#train the single RBM
model = single.rbm(hidden = 30, data=data, learning_rate=.09, epochs=1000, batch_size=100, momentum=0.9)

par(mfrow=c(1,2))
id <- 57

#Original
original <- matrix(data[ ,id], byrow = T, ncol = 16)
original <- t(original[16:1, 1:16]) #correct orientation
image(original, col = grey(seq(0, 1, 0.001)), main = 'Original')

#visible to hidden
features <- up.rbm(model, data)

#hidden back to visible
reconstructions <- down.rbm(model, features)

#Reconstruction
reconstruction <- matrix(reconstructions[id, ], byrow = T, ncol = 16)
reconstruction <- t(reconstruction[16:1, 1:16]) #correct orientation
image(reconstruction, col = grey(seq(0, 1, 0.001)), main = 'Reconstruction')
}

