% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackRBM.R
\name{StackRBM}
\alias{StackRBM}
\title{Stacked Restricted Boltzmann Machine}
\usage{
StackRBM(x, y, n.iter = 100, layers = c(100, 100, 30),
  learning.rate = 0.1, size.minibatch = 10, lambda = 0.1,
  momentum = 0.5)
}
\arguments{
\item{x}{A matrix with binary features of shape samples * features.}

\item{y}{A matrix with labels for the data, only when the last layer is a classification RBM. (Optional)}

\item{n.iter}{Number of epochs for training each RBM layer.}

\item{layers}{Vector with the number of hidden nodes for each RBM layer.}

\item{learning.rate}{The learning rate for training each RBM layer.}

\item{size.minibatch}{The size of the minibatches used for training.}

\item{lambda}{The sparsity penalty lambda to prevent the system from overfitting.}

\item{momentum}{Speeds up the gradient descent learning.}
}
\value{
A list with the trained weights of the stacked RBM that can be used for the predict RBM function when a classification RBM is at
the top layeror the ReconstructRBM function to reconstruct data with the model.
}
\description{
Function to stack several Restricted Boltzmann Machines, trained greedily by training a RBM (using the RBM function) at each layer and
then using the output of that RBM to train the next layer RBM.
}
