% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictDBN.R
\name{PredictDBN}
\alias{PredictDBN}
\title{Predict DBN}
\usage{
PredictDBN(test, labels, model, layers)
}
\arguments{
\item{test}{Is the test-data (matrix) on which the user wants to make predictions.}

\item{labels}{Is a matrix with the corresponding labels for test-data.}

\item{model}{Is the trained DBN model.}

\item{layers}{Define the number of layers.}
}
\value{
A list with a confusion matrix and the accuracy of the predictions.
}
\description{
Function for predicting on a test set with a Deep Belief Network (Trained with DBN)
}
