% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbm.R
\name{print.rbm}
\alias{print.rbm}
\title{Print method for a Restricted Boltzmann Machine}
\usage{
\method{print}{rbm}(x, ...)
}
\arguments{
\item{x}{a RBM object}

\item{...}{not used}
}
\description{
This function prints the weights for a RBM
}

