% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rbm.R
\name{rbm}
\alias{rbm}
\title{rbm}
\usage{
rbm(data_array, CUTOFF = 10^(-9))
}
\arguments{
\item{data_array}{the raw data (should be positive)}

\item{CUTOFF}{the acceptable rounding error}
}
\description{
Computes the RBM estimator along the tuning path.
}

