% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReconstructRBM.R
\name{ReconstructRBM}
\alias{ReconstructRBM}
\title{Reconstruct RBM}
\usage{
ReconstructRBM(test, model, layers = 1)
}
\arguments{
\item{test}{A test example to reconstruct, works only with one example.}

\item{model}{A model trained with either the RBM or StackRBM function.}

\item{layers}{The number of layers that was used for training the StackRBM function.}
}
\value{
A plot of the original and reconstructed image (model) side by side.
}
\description{
Reconstructs an input image with a trained RBM or stacked RBM.
}
\examples{
# Load MNIST data
data(MNIST)

# Fit unsupervised RBM 
mod <- RBM(MNIST$trainX, n.iter = 1000, n.hidden = 100)

# Reconstruct a training image
ReconstructRBM(MNIST$testX[6, ], model = mod)


}
