% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stacked_rbm.R
\name{stacked_rbm}
\alias{stacked_rbm}
\title{Fit a Stack of Restricted Boltzmann Machines}
\usage{
stacked_rbm(x, layers = c(30, 100, 30), learning_rate = 0.1,
  verbose_stack = TRUE, use_gpu = FALSE, ...)
}
\arguments{
\item{x}{a sparse matrix}

\item{layers}{an integer vector of the number of neurons in each RBM}

\item{learning_rate}{The learning rate}

\item{verbose_stack}{Print messages while training the stack}

\item{use_gpu}{use rbm_gpu instead of rbm to train the rbm on a gpu}

\item{...}{passed to the rbm function}
}
\value{
a stacked_rbm object
}
\description{
Fit a Stack of Restricted Boltzmann Machines
}
\examples{
#Setup a dataset
set.seed(10)
print('Data from: https://github.com/echen/restricted-boltzmann-machines')
#Big SF/fantasy fan.
Alice <- c('Harry_Potter' = 1, Avatar = 1, 'LOTR3' = 1, Gladiator = 0, Titanic = 0, Glitter = 0)
#SF/fantasy fan, but doesn't like Avatar.
Bob <- c('Harry_Potter' = 1, Avatar = 0, 'LOTR3' = 1, Gladiator = 0, Titanic = 0, Glitter = 0)
#Big SF/fantasy fan.
Carol <- c('Harry_Potter' = 1, Avatar = 1, 'LOTR3' = 1, Gladiator = 0, Titanic = 0, Glitter = 0)
#Big Oscar winners fan.
David <- c('Harry_Potter' = 0, Avatar = 0, 'LOTR3' = 1, Gladiator = 1, Titanic = 1, Glitter = 0)
#Oscar winners fan, except for Titanic.
Eric <- c('Harry_Potter' = 0, Avatar = 0, 'LOTR3' = 1, Gladiator = 1, Titanic = 0, Glitter = 0)
#Big Oscar winners fan.
Fred <- c('Harry_Potter' = 0, Avatar = 0, 'LOTR3' = 1, Gladiator = 1, Titanic = 1, Glitter = 0)
dat <- rbind(Alice, Bob, Carol, David, Eric, Fred)

Stacked_RBM <- stacked_rbm(dat)
}
\references{
\itemize{
\item \url{http://blog.echen.me/2011/07/18/introduction-to-restricted-boltzmann-machines}
}
}

