% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stacked_rbm.R
\name{stacked_rbm}
\alias{stacked_rbm}
\title{Fit a Stack of Restricted Boltzmann Machines}
\usage{
stacked_rbm(x, layers = c(30, 100, 30), learning_rate = 0.1,
  verbose_stack = TRUE, use_gpu = FALSE, ...)
}
\arguments{
\item{x}{a sparse matrix}

\item{layers}{an integer vector of the number of neurons in each RBM}

\item{learning_rate}{The learning rate}

\item{verbose_stack}{Print messages while training the stack}

\item{use_gpu}{use rbm_gpu instead of rbm to train the rbm on a gpu}

\item{...}{passed to the rbm function}
}
\value{
a stacked_rbm object
}
\description{
Fit a Stack of Restricted Boltzmann Machines
}
\examples{
#Setup a dataset
set.seed(10)
data(movie_reviews)
data(george_reviews)
Stacked_RBM <- stacked_rbm(movie_reviews, layers = c(3, 3), max_epochs=25)
predict(Stacked_RBM, movie_reviews)
}
\references{
\itemize{
\item \url{http://blog.echen.me/2011/07/18/introduction-to-restricted-boltzmann-machines}
}
}

